+++ src/routes.rs
@@ -10,23 +10,55 @@ pub struct Route {
 }
 
 impl Route {
-    pub fn parse(path: &str, route: &str) -> Route {
-        // path: full path (127.0.0.1:3000)
-        // route: url route (/)
-        // slugs: extracted slugs from the path
-        // (/user/32, /user/<id> -> ["id", "32"])
-        // params: query params (/?q="hi" -> ["q", "hi"])
-        // routes: ["user", "32"]
+    pub fn parse(path: &'static str, route: &'static str) -> Route {
         Self {
             path: String::from(path),
             route: String::from(route),
-            slugs: HashMap::new(),
-            params: HashMap::new(),
-            routes: Vec::new(),
+            slugs: parse_slug(path, route),
+            params: parse_params(path.to_string()),
+            routes: path.split("/").map(String::from).collect(),
         }
     }
-    pub fn stringify(&self) -> String {
-        // converts Route to the String
-        String::from(&self.route)
+}
+
+pub fn parse_params(query: String) -> HashMap<String, String> {
+    let mut params: HashMap<String, String> = HashMap::new();
+    let mut queries = query.split("?");
+    for query in queries.nth(1).unwrap_or("").split("&") {
+        let key = query.split('=').nth(0).unwrap_or("").trim().to_string();
+        let value = query.split('=').nth(1).unwrap_or("").trim().to_string();
+        params.insert(key, value);
     }
+    params
+}
+
+pub fn parse_slug(path: &'static str, pattern: &'static str) -> HashMap<String, String> {
+    let mut slugs = HashMap::new();
+    for (i, route) in pattern.split("/").enumerate() {
+        if route.starts_with("<") && route.ends_with(">") {
+            let slug = route.trim_start_matches("<").trim_end_matches(">");
+            let value = path.split("/").nth(i).unwrap_or("").trim().to_string();
+            slugs.insert(slug.to_string(), value);
+        }
+    }
+    slugs
+}
+
+pub fn match_dynamic(path: String, route: String) -> bool {
+    let path_routes = path.split("/").collect::<Vec<&str>>();
+    let route_routes = route.split("/").collect::<Vec<&str>>();
+
+    if path_routes.len() != route_routes.len() {
+        return false;
+    }
+
+    for (i, route_route) in route_routes.iter().enumerate() {
+        if route_route.starts_with("<") && route_route.ends_with(">") {
+            continue;
+        } else if *route_route != path_routes[i] {
+            return false;
+        }
+    }
+
+    true
 }