+++ changelog.md
@@ -1,11 +1,17 @@
 # Changelog
 
+### 0.0.3
+
+- added dynamic routing
+- fixed request `body` and `headers` not passing in `req`
+- fixed request parsing
+- fixed route parsing
+
 ### 0.0.2
 
 - added `req` and `res` arguments for request methods
 - added `POST`, `PUT`, `PATCH`, `DELETE`
 - updated method syntax
-- 
 
 ### 0.0.1